# atlas-core/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/menu/delete/{id}:
        delete:
            tags:
                - SysMenu
            operationId: SysMenu_DeleteSysMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.EmptyReply'
    /v1/menu/info/{id}:
        get:
            tags:
                - SysMenu
            operationId: SysMenu_GetSysMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.GetSysMenuReply'
    /v1/menu/list:
        post:
            tags:
                - SysMenu
            operationId: SysMenu_ListSysMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.ListSysMenuRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.ListSysMenuReply'
    /v1/menu/menuTree/{roleId}:
        get:
            tags:
                - SysMenu
            operationId: SysMenu_GetSysRoleMenu
            parameters:
                - name: roleId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.GetSysRoleMenuReply'
    /v1/menu/save:
        post:
            tags:
                - SysMenu
            operationId: SysMenu_CreateSysMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.SysMenuRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.EmptyReply'
...
    /v1/menu/update:
        put:
            tags:
                - SysMenu
            operationId: SysMenu_UpdateSysMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.SysMenuRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.EmptyReply'
...
components:
    schemas:
...
        api.core.v1.EmptyReply:
            type: object
            properties: {}
...
        api.core.v1.GetSysMenuReply:
            type: object
            properties:
                menu:
                    $ref: '#/components/schemas/api.core.v1.MenuReply'
...
        api.core.v1.GetSysRoleMenuReply:
            type: object
            properties:
                menus:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.RoleMenuReply'
                checkedKeys:
                    type: array
                    items:
                        type: string
...
        api.core.v1.ListSysMenuRep:
            type: object
            properties: {}
        api.core.v1.ListSysMenuReply:
            type: object
            properties:
                menus:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.MenuReply'
...
        api.core.v1.MenuReply:
            type: object
            properties:
                createdAt:
                    type: string
                remark:
                    type: string
                menuId:
                    type: string
                menuName:
                    type: string
                parentName:
                    type: string
                parentId:
                    type: string
                orderNum:
                    type: string
                path:
                    type: string
                component:
                    type: string
                query:
                    type: string
                isFrame:
                    type: string
                isCache:
                    type: string
                menuType:
                    type: string
                visible:
                    type: string
                status:
                    type: string
                perms:
                    type: string
                icon:
                    type: string
                children:
                    type: array
                    items:
                        type: string
...
        api.core.v1.RoleMenuReply:
            type: object
            properties:
                id:
                    type: string
                pid:
                    type: string
                label:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.RoleMenuReply'
...
        api.core.v1.SysMenuRep:
            type: object
            properties:
                menuId:
                    type: string
                parentId:
                    type: string
                menuName:
                    type: string
                icon:
                    type: string
                menuType:
                    type: string
                orderNum:
                    type: string
                isFrame:
                    type: string
                isCache:
                    type: string
                visible:
                    type: string
                status:
                    type: string
                path:
                    type: string
                component:
                    type: string
                perms:
                    type: string
                query:
                    type: string
...